backend/requirements.txt
opencv-python
numpy
deepface
pillow
tf-keras
